abstract display config and adjust order
