+++ frontend/src/scenes/insights/Insights.scss
@@ -9,12 +9,6 @@
         .ant-btn-loading-icon {
             display: none;
         }
-
-        &.horizontal-ui {
-            .ant-tabs-tab {
-                padding: 0; // More compact tabs to save vertical space
-            }
-        }
     }
     hr {
         margin: 1rem 0;
@@ -32,12 +26,43 @@
         }
     }
 
-    .horizontal-ui-card {
-        .ant-card-body {
-            padding: $default_spacing * 0.6 $default_spacing;
+    .insight-controls {
+        overflow: visible;
+    }
+
+    .insights-graph-container {
+        .ant-card-head {
+            background-color: rgba(0, 0, 0, 0.03);
         }
-        margin-bottom: $default_spacing / 2;
-        border: 1px solid $border_light;
+    }
+
+    &.horizontal-ui {
+        .top-bar {
+            .ant-tabs-tab {
+                padding: 0; // More compact tabs to save vertical space
+            }
+        }
+
+        .insight-controls {
+            margin-bottom: $default_spacing / 2;
+            border: 1px solid $border_light;
+            .ant-card-body {
+                padding: $default_spacing * 0.6 $default_spacing;
+            }
+        }
+
+        .insights-graph-container {
+            .ant-card-head {
+                min-height: unset;
+                @extend .mixin-elevated;
+                background-color: rgba(0, 0, 0, 0.02);
+                border: 1px solid $border_light;
+
+                .ant-card-head-title {
+                    padding: $default_spacing / 2 0;
+                }
+            }
+        }
     }
 }
