+++ frontend/src/scenes/insights/Insights.tsx
@@ -2,24 +2,13 @@ import React, { useState } from 'react'
 import { useActions, useMountedLogic, useValues, BindLogic } from 'kea'
 
 import { isMobile, Loading } from 'lib/utils'
-import { SaveToDashboard } from 'lib/components/SaveToDashboard/SaveToDashboard'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import { DateFilter } from './DateFilter'
-import { IntervalFilter } from 'lib/components/IntervalFilter/IntervalFilter'
 
 import { PageHeader } from 'lib/components/PageHeader'
 
-import { ChartFilter } from 'lib/components/ChartFilter'
 import { Tabs, Row, Col, Card, Button, Tooltip } from 'antd'
-import {
-    ACTIONS_LINE_GRAPH_LINEAR,
-    ACTIONS_LINE_GRAPH_CUMULATIVE,
-    ACTIONS_TABLE,
-    ACTIONS_PIE_CHART,
-    ACTIONS_BAR_CHART_VALUE,
-    FUNNEL_VIZ,
-} from 'lib/constants'
+import { ACTIONS_LINE_GRAPH_LINEAR, ACTIONS_LINE_GRAPH_CUMULATIVE, FUNNEL_VIZ } from 'lib/constants'
 import { annotationsLogic } from '~/lib/components/Annotations'
 import { router } from 'kea-router'
 
@@ -31,7 +20,6 @@ import { RetentionTab, SessionTab, TrendTab, PathTab, FunnelTab } from './Insigh
 import { FunnelViz } from 'scenes/funnels/FunnelViz'
 import { funnelLogic } from 'scenes/funnels/funnelLogic'
 import { insightLogic, logicFromInsight, ViewType } from './insightLogic'
-import { CompareFilter } from 'lib/components/CompareFilter/CompareFilter'
 import { InsightHistoryPanel } from './InsightHistoryPanel'
 import { SavedFunnels } from './SavedCard'
 import { ReloadOutlined } from '@ant-design/icons'
@@ -44,67 +32,14 @@ import { People } from 'scenes/funnels/People'
 import { TrendLegend } from './TrendLegend'
 import { TrendInsight } from 'scenes/trends/Trends'
 import { trendsLogic } from 'scenes/trends/trendsLogic'
-import { TZIndicator } from 'lib/components/TimezoneAware'
-import { DashboardItemType, DisplayType, FilterType, HotKeys } from '~/types'
+import { DashboardItemType, HotKeys } from '~/types'
 import { useKeyboardHotkeys } from 'lib/hooks/useKeyboardHotkeys'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { DashboardItemHeader } from 'scenes/dashboard/DashboardItemHeader'
+import { InsightDisplayConfig } from './InsightTabs/InsightDisplayConfig'
 
 dayjs.extend(relativeTime)
 const { TabPane } = Tabs
-
-const showIntervalFilter = function (activeView: ViewType, filter: FilterType): boolean {
-    switch (activeView) {
-        case ViewType.FUNNELS:
-            return filter.display === ACTIONS_LINE_GRAPH_LINEAR
-        case ViewType.RETENTION:
-        case ViewType.PATHS:
-            return false
-        case ViewType.TRENDS:
-        case ViewType.STICKINESS:
-        case ViewType.LIFECYCLE:
-        case ViewType.SESSIONS:
-        default:
-            return ![ACTIONS_PIE_CHART, ACTIONS_TABLE, ACTIONS_BAR_CHART_VALUE].includes(filter.display || '') // sometimes insights aren't set for trends
-    }
-}
-
-const showChartFilter = function (activeView: ViewType, featureFlags: Record<string, boolean>): boolean {
-    switch (activeView) {
-        case ViewType.TRENDS:
-        case ViewType.STICKINESS:
-        case ViewType.SESSIONS:
-        case ViewType.RETENTION:
-            return true
-        case ViewType.FUNNELS:
-            return featureFlags['funnel-trends-1269']
-        case ViewType.LIFECYCLE:
-        case ViewType.PATHS:
-            return false
-        default:
-            return true // sometimes insights aren't set for trends
-    }
-}
-
-const showDateFilter = {
-    [`${ViewType.TRENDS}`]: true,
-    [`${ViewType.STICKINESS}`]: true,
-    [`${ViewType.LIFECYCLE}`]: true,
-    [`${ViewType.SESSIONS}`]: true,
-    [`${ViewType.FUNNELS}`]: true,
-    [`${ViewType.RETENTION}`]: false,
-    [`${ViewType.PATHS}`]: true,
-}
-
-const showComparePrevious = {
-    [`${ViewType.TRENDS}`]: true,
-    [`${ViewType.STICKINESS}`]: true,
-    [`${ViewType.LIFECYCLE}`]: false,
-    [`${ViewType.SESSIONS}`]: true,
-    [`${ViewType.FUNNELS}`]: false,
-    [`${ViewType.RETENTION}`]: false,
-    [`${ViewType.PATHS}`]: false,
-}
 
 function InsightHotkey({ hotkey }: { hotkey: HotKeys }): JSX.Element {
     /* Temporary element to only show hotkeys when feature flag is active */
@@ -131,7 +66,8 @@ export function Insights(): JSX.Element {
     const { reportHotkeyNavigation } = useActions(eventUsageLogic)
 
     const { loadResults } = useActions(logicFromInsight(activeView, { dashboardItemId: null, filters: allFilters }))
-    const dateFilterDisabled = activeView === ViewType.FUNNELS && isFunnelEmpty(allFilters)
+
+    const horizontalUI = featureFlags['4050-query-ui-optB']
 
     const handleHotkeyNavigation = (view: ViewType, hotkey: HotKeys): void => {
         setActiveView(view)
@@ -167,7 +103,7 @@ export function Insights(): JSX.Element {
     )
 
     return (
-        <div className="insights-page">
+        <div className={`insights-page${horizontalUI ? ' horizontal-ui' : ''}`}>
             <Header dashboardItem={dashboardItem} />
             {!dashboardItem && (
                 <Row justify="space-between" align="middle" className="top-bar">
@@ -176,7 +112,7 @@ export function Insights(): JSX.Element {
                         style={{
                             overflow: 'visible',
                         }}
-                        className={`top-bar${featureFlags['4050-query-ui-optB'] ? ' horizontal-ui' : ''}`}
+                        className="top-bar"
                         onChange={(key) => setActiveView(key)}
                         animated={false}
                         tabBarExtraContent={{
@@ -291,11 +227,8 @@ export function Insights(): JSX.Element {
                     </Col>
                 ) : (
                     <>
-                        <Col xs={24} xl={featureFlags['4050-query-ui-optB'] ? 24 : 7}>
-                            <Card
-                                className={featureFlags['4050-query-ui-optB'] ? 'horizontal-ui-card' : ''}
-                                style={{ overflow: 'visible' }}
-                            >
+                        <Col xs={24} xl={horizontalUI ? 24 : 7}>
+                            <Card className="insight-controls">
                                 <div>
                                     {/*
                                 These are insight specific filters.
@@ -323,58 +256,23 @@ export function Insights(): JSX.Element {
                                 </Card>
                             )}
                         </Col>
-                        <Col xs={24} xl={featureFlags['4050-query-ui-optB'] ? 24 : 17}>
-                            {' '}
+                        <Col xs={24} xl={horizontalUI ? 24 : 17}>
                             {/* TODO: extract to own file. Props: activeView, allFilters, showDateFilter, dateFilterDisabled, annotationsToCreate; lastRefresh, showErrorMessage, showTimeoutMessage, isLoading; ... */}
-                            {/*
-                        These are filters that are reused between insight features.
-                        They each have generic logic that updates the url
-                        */}
+                            {/* These are filters that are reused between insight features. They
+                                each have generic logic that updates the url
+                            */}
                             <Card
                                 title={
-                                    <div style={{ display: 'flex', alignItems: 'center' }}>
-                                        <TZIndicator style={{ float: 'left' }} />
-                                        <div style={{ width: '100%', textAlign: 'right' }}>
-                                            {showIntervalFilter(activeView, allFilters) && (
-                                                <IntervalFilter view={activeView} />
-                                            )}
-                                            {showChartFilter(activeView, featureFlags) && (
-                                                <ChartFilter
-                                                    onChange={(display: DisplayType) => {
-                                                        if (
-                                                            display === ACTIONS_TABLE ||
-                                                            display === ACTIONS_PIE_CHART
-                                                        ) {
-                                                            clearAnnotationsToCreate()
-                                                        }
-                                                    }}
-                                                    filters={allFilters}
-                                                    disabled={allFilters.insight === ViewType.LIFECYCLE}
-                                                />
-                                            )}
-
-                                            {showDateFilter[activeView] && (
-                                                <DateFilter
-                                                    defaultValue="Last 7 days"
-                                                    disabled={dateFilterDisabled}
-                                                    bordered={false}
-                                                />
-                                            )}
-
-                                            {showComparePrevious[activeView] && <CompareFilter />}
-                                            <SaveToDashboard
-                                                item={{
-                                                    entity: {
-                                                        filters: allFilters,
-                                                        annotations: annotationsToCreate,
-                                                    },
-                                                }}
-                                            />
-                                        </div>
-                                    </div>
+                                    <InsightDisplayConfig
+                                        activeView={activeView}
+                                        allFilters={allFilters}
+                                        annotationsToCreate={annotationsToCreate}
+                                        clearAnnotationsToCreate={clearAnnotationsToCreate}
+                                        horizontalUI={horizontalUI}
+                                    />
                                 }
-                                headStyle={{ backgroundColor: 'rgba(0,0,0,.03)' }}
                                 data-attr="insights-graph"
+                                className="insights-graph-container"
                             >
                                 <div>
                                     {lastRefresh && (
@@ -448,10 +346,6 @@ export function Insights(): JSX.Element {
             </Row>
         </div>
     )
-}
-
-const isFunnelEmpty = (filters: FilterType): boolean => {
-    return (!filters.actions && !filters.events) || (filters.actions?.length === 0 && filters.events?.length === 0)
 }
 
 function FunnelInsight(): JSX.Element {
